requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-chess==1.999
psycopg2-binary==2.9.9
gunicorn==21.2.0
